+++ R/method_baseline.R
@@ -17,6 +17,8 @@
 #'  \item{\code{hamming-loss}}{Predict the labels that are associated with more
 #'   than 50\% of instances.}
 #'  \item{\code{subset-accuracy}}{Predict the most common labelset.}
+#'  \item{\code{ranking-loss}}{Predict a ranking based on the most frequent
+#'   labels.}
 #' }
 #'
 #' @param mdata A mldr dataset used to train the binary models.
@@ -47,7 +49,7 @@
 #' model <- baseline(toyml, "F1")
 #' model <- baseline(toyml, "subset-accuracy")
 baseline <- function (mdata, metric = c("general", "F1", "hamming-loss",
-                      "subset-accuracy"), ...) {
+                      "subset-accuracy", "ranking-loss"), ...) {
   # Validations
   if (class(mdata) != "mldr") {
     stop("First argument must be an mldr object")
@@ -82,6 +84,11 @@ baseline <- function (mdata, metric = c("general", "F1", "hamming-loss",
     "subset-accuracy" = function (mdata){
       lbl <- as.numeric(unlist(strsplit(names(which.max(mdata$labelsets)), "")))
       rownames(mdata$labels)[lbl == 1]
+    },
+    "ranking-loss" = function (mdata) {
+      rk <- order(mdata$labels$freq, decreasing=TRUE)
+      half <- mdata$labels$freq / 2
+      half + (0.49 - max(half[-rk[seq(round(mdata$measures$cardinality,0))]]))
     }
   )
 
@@ -121,10 +128,16 @@ predict.BASELINEmodel <- function (object, newdata,
   }
 
   newdata <- utiml_newdata(newdata)
-  prediction <- matrix(0, nrow=nrow(newdata), ncol=length(object$labels),
-                       dimnames = list(rownames(newdata), object$labels))
 
-  prediction[, object$predict] <- 1
+  if (mode(object$predict) == "numeric") {
+    prediction <- matrix(rep(object$predict, nrow(newdata)), byrow = TRUE,
+                         nrow=nrow(newdata), ncol=length(object$labels),
+                         dimnames = list(rownames(newdata), object$labels))
+  } else {
+    prediction <- matrix(0, nrow=nrow(newdata), ncol=length(object$labels),
+                         dimnames = list(rownames(newdata), object$labels))
+    prediction[, object$predict] <- 1
+  }
 
   as.mlresult(prediction, probability = probability)
 }