+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ### New Features
 * MLKNN algorithm
+* ranking-loss baseline
 
 ### Bug fixes
 * Stratification sampling to support instances without labels